refactor: remove repeated files; make symbolic links
