+++ egs/vcc2020/tts1_en_de/local/text
@@ -1,0 +1,1 @@
+../../../ljspeech/tts1/local/text
