+++ egs/vcc2020/tts1_en_fi/local/clean_text_css10.py
@@ -9,8 +9,9 @@ import codecs
 import json
 import os
 
-from text.cleaners import custom_english_cleaners
-import nltk
+from text.cleaners import (lowercase, expand_numbers, expand_abbreviations,
+                           expand_symbols, remove_unnecessary_symbols,
+                           uppercase, collapse_whitespace)
 
 try:
     # For phoneme conversion, use https://github.com/Kyubyong/g2p.
@@ -21,9 +22,7 @@ except ImportError:
     raise ImportError("g2p_en is not installed. please run `. ./path.sh && pip install g2p_en`.")
 except LookupError:
     # NOTE: we need to download dict in initial running
-
     import ssl
-
     try:
         _create_unverified_https_context = ssl._create_unverified_context
     except AttributeError:
@@ -32,11 +31,15 @@ except LookupError:
         ssl._create_default_https_context = _create_unverified_https_context
     nltk.download("punkt")
 
-
-def g2p(text):
-    """Convert grapheme to phoneme."""
-    tokens = filter(lambda s: s != " ", f_g2p(text))
-    return ' '.join(tokens)
+def custom_finnish_cleaners(text):
+    text = lowercase(text)
+    text = expand_numbers(text)
+    text = expand_abbreviations(text)
+    text = expand_symbols(text)
+    text = remove_unnecessary_symbols(text)
+    text = uppercase(text)
+    text = collapse_whitespace(text)
+    return text
 
 def main():
     parser = argparse.ArgumentParser()
@@ -44,8 +47,8 @@ def main():
                         help="language tag (can be used for multi lingual case)")
     parser.add_argument("--spk_tag", type=str,
                         help="speaker tag")
-    parser.add_argument("jsons", nargs="+", type=str,
-                        help="*_mls.json filenames")
+    parser.add_argument("transcription", type=str,
+                        help="transcription filename")
     parser.add_argument("out", type=str,
                         help="output filename")
     parser.add_argument("trans_type", type=str, default="char",
@@ -57,15 +60,13 @@ def main():
         os.makedirs(dirname)
 
     with codecs.open(args.out, "w", encoding="utf-8") as out:
-        for filename in sorted(args.jsons):
-            with codecs.open(filename, "r", encoding="utf-8") as f:
-                js = json.load(f)
-            for key in sorted(js.keys()):
-                uid = args.spk_tag + "_" + key.replace(".wav", "")
-                content = js[key]["clean"]
-                text = custom_english_cleaners(content.rstrip())
+        with codecs.open(args.transcription, "r", encoding="utf-8") as f:
+            for line in f.read().splitlines():
+                path, _, content, _ = line.split('|')
+                uid = args.spk_tag + "_" + os.path.basename(path).replace(".wav", "")
+                text = custom_finnish_cleaners(content.rstrip())
                 if args.lang_tag is None:
-                    line = "%s %s \n" % (uid, text)
+                    line = "%s %s\n" % (uid, text)
                 else:
                     line = "%s <%s> %s\n" % (uid, args.lang_tag, text)
                 out.write(line)
