+++ egs/vcc2020/tts1_en_fi/local/clean_text_mailabs.py
@@ -1,0 +1,1 @@
+../../tts1_en_zh/local/clean_text_mailabs.py
