+++ egs/vcc2020/tts1_en_fi/local/text
@@ -1,0 +1,1 @@
+../../../ljspeech/tts1/local/text
